ARC model

Arcs are unique paths that connect nodes. In a polygon layer with shared boundaries, the arcs are the linear features that have no branches.

ARC(x, ...)

# S3 method for default
ARC(x, ...)

# S3 method for PATH
ARC(x, ...)

Arguments

x

input model

...

arguments passed to methods

Value

ARC model

Details

Nodes are the vertices where three or more arcs meet. An arc can exist without including any nodes, a path that has no neighbouring relationship with another path.

This is not the same terminology as used by other systems, such as "arc-node". The arc_link_vertex mapping is inherently ordered, but we don't consider order of arcs. Duplicated arcs (i.e. complementary turns around neighbouring polygons) are not kept. The object_link_arc mapping records which arc belongs to the objects, so feature polygons can in theory be reconstructed within objects by tracing arc_link_vertex start and end point identity.

Examples

a <- ARC(minimal_mesh)
sc_arc(a)
#> # A tibble: 4 × 2
#>   arc_   ncoords_
#>   <chr>     <int>
#> 1 0Yofvz        4
#> 2 LC4Gr6        6
#> 3 gkG9CQ        7
#> 4 mqDWYw        4
sc_arc(minimal_mesh)
#> # A tibble: 4 × 2
#>   arc_   ncoords_
#>   <chr>     <int>
#> 1 Ei1xSg        4
#> 2 JO6XIJ        6
#> 3 TiuBVr        7
#> 4 a9BdiJ        4