Arc-node topology.

Return a label and vertex count of each arc.

sc_arc(x, ...)

# S3 method for default
sc_arc(x, ...)

# S3 method for ARC
sc_arc(x, ...)

Arguments

x
input object

...
arguments for methods

Details

Arcs are unbranched paths within the line segment graph. Nodes are the vertices where three or more arcs meet.

As with the PATH and SC models the arc id values will only be relevant when those entities are identified and labelled. Running sc_arc on a simple features model (for example) will identify them and return a summary, but without having any record of what they refer to. Use ARC(x) first to work with models that don't included labels.

Examples

sc_arc(minimal_mesh)

#> # A tibble: 4 x 2
#>   arc_     ncoords_
#>   <chr>       <int>
#> 1 62a3a7ba        2
#> 2 7eb41893        7
#> 3 7ec8785c        6
#> 4 adbd189c        4

ARC(minimal_mesh)[["arc"]]

#> NULL

arc <- ARC(minimal_mesh)
plot(arc)

#> Error in plot.new(): figure margins too large

points(arc$vertex[match(sc_node(arc)$vertex_, arc$vertex$vertex_), c("x_", "y_")])

#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet

arc <- ARC(polymesh)
plot(arc)

#> Error in plot.xy(xy, type, ...): invalid graphics state

title("arcs and nodes")

#> Error in title("arcs and nodes"): invalid graphics state

points(arc$vertex[match(sc_node(arc)$vertex_, arc$vertex$vertex_), c("x_", "y_")])

#> Error in plot.xy(xy.coords(x, y), type = type, ...): invalid graphics state