Path decomposition

Start in the middle, and build the 'path-link-vertex' table.

Paths.

sc_path(x, ...)

# S3 method for list
sc_path(x, ids = NULL, ...)

# S3 method for default
sc_path(x, ...)

# S3 method for PATH
sc_path(x, ...)

# S3 method for sfc_TIN
sc_path(x, ...)

# S3 method for PATH0
sc_path(x, ...)

# S3 method for ARC
sc_path(x, ...)

# S3 method for SC
sc_path(x, ...)

# S3 method for SC0
sc_path(x, ...)

# S3 method for matrix
sc_path(x, ...)

# S3 method for sf
sc_path(x, ids = NULL, ...)

# S3 method for sfc
sc_path(x, ids = NULL, ...)

# S3 method for MULTIPOLYGON
sc_path(x, ...)

# S3 method for POLYGON
sc_path(x, ...)

# S3 method for LINESTRING
sc_path(x, ...)

# S3 method for MULTILINESTRING
sc_path(x, ...)

# S3 method for POINT
sc_path(x, ...)

# S3 method for MULTIPOINT
sc_path(x, ...)

# S3 method for GEOMETRYCOLLECTION
sc_path(x, ...)

# S3 method for Spatial
sc_path(x, ids = NULL, ...)

Arguments

x
input object

...
arguments passed to methods

ids
object id, one for each object in the sfc

Value

data frame of path identity and properties

Details

Paths have properties of their type, their number of vertices, their geometric dimension and which object they occur in.

See also

Examples

sc_path(minimal_mesh)

#> # A tibble: 3 x 6
#>    ncol type         subobject object_ path_  ncoords_
#>   <int> <chr>            <int> <chr>   <chr>     <int>
#> 1     2 MULTIPOLYGON         1 o4bU5B  nef98r        8
#> 2     2 MULTIPOLYGON         1 o4bU5B  9MHhmS        6
#> 3     2 MULTIPOLYGON         1 EqlQz8  zdqaBl        5

sc_path(PATH(minimal_mesh))

#> # A tibble: 3 x 6
#>    ncol type         subobject object_ path_  ncoords_
#>   <int> <chr>            <int> <chr>   <chr>     <int>
#> 1     2 MULTIPOLYGON         1 I07VjN  XQ6fYL        8
#> 2     2 MULTIPOLYGON         1 I07VjN  grPIJs        6
#> 3     2 MULTIPOLYGON         1 Ih3gM4  1DISJK        5

sc_path(sfzoo$multipolygon)

#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 SYdZ3o
#> 2     4     2 MULTIPOLYGON         1 xYtgnG
#> 3     5     2 MULTIPOLYGON         2 sgSKvc
#> 4     5     2 MULTIPOLYGON         2 8wuzbO
#> 5     4     2 MULTIPOLYGON         3 yMG8Qz

sc_path(sfzoo$polygon)

#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON 19zhuE
#> 2     4     2 POLYGON u1s4A1

sc_path(sfzoo$linestring)

#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING 8M4Ftw

sc_path(sfzoo$multilinestring)

#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING wtD7OH
#> 2     4     2 MULTILINESTRING cdSXtn
#> 3     2     2 MULTILINESTRING NXcVFr

sc_path(sfzoo$point)

#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT uuH0Pz

sc_path(sfzoo$multipoint)

#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT W5fKkR

sc_path(sfzoo$multipoint)

#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT A7a2K9