Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     1     2
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1   3.2   4  
#> 2   3     4.6
#> 3   3.8   4.4
#> 4   3.5   3.8
#> 5   3.4   3.6
#> 6   3.9   4.5
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     0     3
#> 2     0     4
#> 3     1     5
#> 4     2     5
#> 
#> $multilinestring
#>     x_  y_
#> 1  0.0 3.0
#> 2  0.0 4.0
#> 3  1.0 5.0
#> 4  2.0 5.0
#> 5  0.2 3.0
#> 6  0.2 4.0
#> 7  1.0 4.8
#> 8  2.0 4.8
#> 9  0.0 4.4
#> 10 0.6 5.0
#> 
#> $polygon
#>    x_ y_
#> 1   0  0
#> 2   1  0
#> 3   3  2
#> 4   2  4
#> 5   1  4
#> 6   0  0
#> 7   1  1
#> 8   1  2
#> 9   2  2
#> 10  1  1
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> # … with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT Mgut9n
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT mPxJaP
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING ZYnP5t
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING jLAngr
#> 2     4     2 MULTILINESTRING OPGtNv
#> 3     2     2 MULTILINESTRING JospMa
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON dCfm0L
#> 2     4     2 POLYGON A22DWq
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 oPUPXn
#> 2     4     2 MULTIPOLYGON         1 aBtyrZ
#> 3     5     2 MULTIPOLYGON         2 C515Co
#> 4     5     2 MULTIPOLYGON         2 inGXvc
#> 5     4     2 MULTIPOLYGON         3 3sL0Ef
#> 

## unsure how useful this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT ZPqxR9
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT N0fyDp
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING qzs3nO
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING tb8SIj
#> 2     4     2 MULTILINESTRING m4BUGX
#> 3     2     2 MULTILINESTRING 4DhIgg
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON ZzTsFV
#> 2     4     2 POLYGON 5GQwXq
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 EWMvQs
#> 2     4     2 MULTIPOLYGON         1 XGrsnY
#> 3     5     2 MULTIPOLYGON         2 0tey6Y
#> 4     5     2 MULTIPOLYGON         2 3RWciR
#> 5     4     2 MULTIPOLYGON         3 h994OC
#>