Flight tracks

A data set flight tracks in XYZM form, a form of 4D tracks. Primarily to explore the use of silicate as able to represent this topologically, and to experiment with auto-time-based plotting in anglr.

Details

Provided by Kent Johnson (kent37) in a github discusion where the data was attached in a zip file.

Original form (in extdata/flight_tracks) is a XYZM LINESTRING shapefile containing 144 flight tracks of aircraft departing runway 33L at Boston Logan airport on January 27, 2017. Data is from an ADS-B recorder. Each point includes lat, lon, altitude in feet and time in North American Eastern Standard Time (EST).

Converted via sf into silicate::PATH normal form, see (data-raw/flight_tracks.R).