Path decomposition

Start in the middle, and build the 'path-link-vertex' table.

Paths.

sc_path(x, ...)

# S3 method for list
sc_path(x, ids = NULL, ...)

# S3 method for default
sc_path(x, ...)

# S3 method for PATH
sc_path(x, ...)

# S3 method for sfc_TIN
sc_path(x, ...)

# S3 method for PATH0
sc_path(x, ...)

# S3 method for ARC
sc_path(x, ...)

# S3 method for SC
sc_path(x, ...)

# S3 method for SC0
sc_path(x, ...)

# S3 method for matrix
sc_path(x, ...)

# S3 method for sf
sc_path(x, ids = NULL, ...)

# S3 method for sfc
sc_path(x, ids = NULL, ...)

# S3 method for MULTIPOLYGON
sc_path(x, ...)

# S3 method for POLYGON
sc_path(x, ...)

# S3 method for LINESTRING
sc_path(x, ...)

# S3 method for MULTILINESTRING
sc_path(x, ...)

# S3 method for POINT
sc_path(x, ...)

# S3 method for MULTIPOINT
sc_path(x, ...)

# S3 method for GEOMETRYCOLLECTION
sc_path(x, ...)

# S3 method for Spatial
sc_path(x, ids = NULL, ...)

Arguments

x

input object

...

arguments passed to methods

ids

object id, one for each object in the sfc

Value

data frame of path identity and properties

Details

Paths have properties of their type, their number of vertices, their geometric dimension and which object they occur in.

See also

Examples

sc_path(minimal_mesh)
#> # A tibble: 3 × 6
#>    ncol type         subobject object_ path_  ncoords_
#>   <int> <chr>            <int> <chr>   <chr>     <int>
#> 1     2 MULTIPOLYGON         1 dAeyks  SohCRt        8
#> 2     2 MULTIPOLYGON         1 dAeyks  BeqMEX        6
#> 3     2 MULTIPOLYGON         1 eXEZrL  jvO12C        5
sc_path(PATH(minimal_mesh))
#> # A tibble: 3 × 6
#>    ncol type         subobject object_ path_  ncoords_
#>   <int> <chr>            <int> <chr>   <chr>     <int>
#> 1     2 MULTIPOLYGON         1 HsI7Zd  fBs0dg        8
#> 2     2 MULTIPOLYGON         1 HsI7Zd  1ndq1e        6
#> 3     2 MULTIPOLYGON         1 E7mUDo  osNCUX        5
sc_path(sfzoo$multipolygon)
#> # A tibble: 5 × 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 cghxJb
#> 2     4     2 MULTIPOLYGON         1 qYdf7i
#> 3     5     2 MULTIPOLYGON         2 IWYiQq
#> 4     5     2 MULTIPOLYGON         2 PdmMRD
#> 5     4     2 MULTIPOLYGON         3 zJ7Cy6
sc_path(sfzoo$polygon)
#> # A tibble: 2 × 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON XtuRlZ
#> 2     4     2 POLYGON OcZNwg
sc_path(sfzoo$linestring)
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING EKeDy1
sc_path(sfzoo$multilinestring)
#> # A tibble: 3 × 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING BZxrY0
#> 2     4     2 MULTILINESTRING r5YMA2
#> 3     2     2 MULTILINESTRING FgsKm8
sc_path(sfzoo$point)
#> # A tibble: 1 × 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT jY0IL9
sc_path(sfzoo$multipoint)
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT jcK3Sz
sc_path(sfzoo$multipoint)
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT vSrPJq