Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1  1.00  2.00
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1  3.20  4.00
#> 2  3.00  4.60
#> 3  3.80  4.40
#> 4  3.50  3.80
#> 5  3.40  3.60
#> 6  3.90  4.50
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1  0     3.00
#> 2  0     4.00
#> 3  1.00  5.00
#> 4  2.00  5.00
#> 
#> $multilinestring
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1 0      3.00
#>  2 0      4.00
#>  3 1.00   5.00
#>  4 2.00   5.00
#>  5 0.200  3.00
#>  6 0.200  4.00
#>  7 1.00   4.80
#>  8 2.00   4.80
#>  9 0      4.40
#> 10 0.600  5.00
#> 
#> $polygon
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1  0     0   
#>  2  1.00  0   
#>  3  3.00  2.00
#>  4  2.00  4.00
#>  5  1.00  4.00
#>  6  0     0   
#>  7  1.00  1.00
#>  8  1.00  2.00
#>  9  2.00  2.00
#> 10  1.00  1.00
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1  0     0   
#>  2  1.00  0   
#>  3  3.00  2.00
#>  4  2.00  4.00
#>  5  1.00  4.00
#>  6  0     0   
#>  7  1.00  1.00
#>  8  1.00  2.00
#>  9  2.00  2.00
#> 10  1.00  1.00
#> # ... with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_   
#>   <dbl> <dbl> <chr> <chr>   
#> 1  1.00  2.00 POINT 242a338d
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_   
#>   <int> <int> <chr>      <chr>   
#> 1     6     2 MULTIPOINT 4d19dc35
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_   
#>   <int> <int> <chr>      <chr>   
#> 1     4     2 LINESTRING a3c921a6
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_   
#>   <int> <int> <chr>           <chr>   
#> 1     4     2 MULTILINESTRING bdddeb8d
#> 2     4     2 MULTILINESTRING 5de654d9
#> 3     2     2 MULTILINESTRING 27d73530
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_   
#>   <int> <int> <chr>   <chr>   
#> 1     6     2 POLYGON a3211beb
#> 2     4     2 POLYGON b5fc9bcd
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_   
#>   <int> <int> <chr>            <int> <chr>   
#> 1     6     2 MULTIPOLYGON         1 a2bce15e
#> 2     4     2 MULTIPOLYGON         1 aef487f3
#> 3     5     2 MULTIPOLYGON         2 470f9c1d
#> 4     5     2 MULTIPOLYGON         2 b2eb6d4c
#> 5     4     2 MULTIPOLYGON         3 7603f19c
#> 

## unsure how usefult this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_   
#>   <dbl> <dbl> <chr> <chr>   
#> 1  1.00  2.00 POINT 5ddc2bb8
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_   
#>   <int> <int> <chr>      <chr>   
#> 1     6     2 MULTIPOINT e8a9e0a7
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_   
#>   <int> <int> <chr>      <chr>   
#> 1     4     2 LINESTRING 391bacf2
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_   
#>   <int> <int> <chr>           <chr>   
#> 1     4     2 MULTILINESTRING 1fda630c
#> 2     4     2 MULTILINESTRING 98fee137
#> 3     2     2 MULTILINESTRING d379e372
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_   
#>   <int> <int> <chr>   <chr>   
#> 1     6     2 POLYGON 62bb36a8
#> 2     4     2 POLYGON 5556055b
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_   
#>   <int> <int> <chr>            <int> <chr>   
#> 1     6     2 MULTIPOLYGON         1 e9c56c60
#> 2     4     2 MULTIPOLYGON         1 92f191db
#> 3     5     2 MULTIPOLYGON         2 cbc73f7d
#> 4     5     2 MULTIPOLYGON         2 8f996d7e
#> 5     4     2 MULTIPOLYGON         3 7eee2920
#>