Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     1     2
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1   3.2   4  
#> 2   3     4.6
#> 3   3.8   4.4
#> 4   3.5   3.8
#> 5   3.4   3.6
#> 6   3.9   4.5
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     0     3
#> 2     0     4
#> 3     1     5
#> 4     2     5
#> 
#> $multilinestring
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1   0     3  
#>  2   0     4  
#>  3   1     5  
#>  4   2     5  
#>  5   0.2   3  
#>  6   0.2   4  
#>  7   1     4.8
#>  8   2     4.8
#>  9   0     4.4
#> 10   0.6   5  
#> 
#> $polygon
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> # ... with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_     
#>   <dbl> <dbl> <chr> <chr>     
#> 1     1     2 POINT 322e445ff5
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     6     2 MULTIPOINT d33a1e1072
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     4     2 LINESTRING 22de9e364d
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_     
#>   <int> <int> <chr>           <chr>     
#> 1     4     2 MULTILINESTRING fb6a0e550e
#> 2     4     2 MULTILINESTRING 1450c04c6f
#> 3     2     2 MULTILINESTRING 2d4d6c928f
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_     
#>   <int> <int> <chr>   <chr>     
#> 1     6     2 POLYGON 38e775768d
#> 2     4     2 POLYGON 3a172ac652
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_     
#>   <int> <int> <chr>            <int> <chr>     
#> 1     6     2 MULTIPOLYGON         1 2831a0ec1f
#> 2     4     2 MULTIPOLYGON         1 a023c35458
#> 3     5     2 MULTIPOLYGON         2 3d138e9f19
#> 4     5     2 MULTIPOLYGON         2 228199ae12
#> 5     4     2 MULTIPOLYGON         3 28a9d68fb7
#> 

## unsure how usefult this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_     
#>   <dbl> <dbl> <chr> <chr>     
#> 1     1     2 POINT 6416fce6f6
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     6     2 MULTIPOINT 7e2ed8cdfb
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     4     2 LINESTRING c3f50d3c46
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_     
#>   <int> <int> <chr>           <chr>     
#> 1     4     2 MULTILINESTRING 7b95e47a14
#> 2     4     2 MULTILINESTRING 671482df50
#> 3     2     2 MULTILINESTRING cb481ef255
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_     
#>   <int> <int> <chr>   <chr>     
#> 1     6     2 POLYGON 34dec053be
#> 2     4     2 POLYGON f6ffbff467
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_     
#>   <int> <int> <chr>            <int> <chr>     
#> 1     6     2 MULTIPOLYGON         1 46eaaa949a
#> 2     4     2 MULTIPOLYGON         1 233d324361
#> 3     5     2 MULTIPOLYGON         2 7b86d78698
#> 4     5     2 MULTIPOLYGON         2 beaeaddd36
#> 5     4     2 MULTIPOLYGON         3 e9c01a1efe
#>