Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     1     2
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1   3.2   4  
#> 2   3     4.6
#> 3   3.8   4.4
#> 4   3.5   3.8
#> 5   3.4   3.6
#> 6   3.9   4.5
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     0     3
#> 2     0     4
#> 3     1     5
#> 4     2     5
#> 
#> $multilinestring
#>     x_  y_
#> 1  0.0 3.0
#> 2  0.0 4.0
#> 3  1.0 5.0
#> 4  2.0 5.0
#> 5  0.2 3.0
#> 6  0.2 4.0
#> 7  1.0 4.8
#> 8  2.0 4.8
#> 9  0.0 4.4
#> 10 0.6 5.0
#> 
#> $polygon
#>    x_ y_
#> 1   0  0
#> 2   1  0
#> 3   3  2
#> 4   2  4
#> 5   1  4
#> 6   0  0
#> 7   1  1
#> 8   1  2
#> 9   2  2
#> 10  1  1
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> # ... with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT UxzqUL
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT 0RZeG4
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING 0N0c9J
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING CkeOpG
#> 2     4     2 MULTILINESTRING bFr9x8
#> 3     2     2 MULTILINESTRING GSJh23
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON wtcmjY
#> 2     4     2 POLYGON Ox0S6F
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 imkcvm
#> 2     4     2 MULTIPOLYGON         1 BtJZ9A
#> 3     5     2 MULTIPOLYGON         2 IzLxMi
#> 4     5     2 MULTIPOLYGON         2 e5HBT7
#> 5     4     2 MULTIPOLYGON         3 cD4ISB
#> 

## unsure how useful this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT l5Ho0n
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT GqwfCT
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING 4YhAKT
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING zjrJzH
#> 2     4     2 MULTILINESTRING AXn7t0
#> 3     2     2 MULTILINESTRING sowPlU
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON aP9Aa1
#> 2     4     2 POLYGON Ssg21c
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 eoPbEO
#> 2     4     2 MULTIPOLYGON         1 ODuXSU
#> 3     5     2 MULTIPOLYGON         2 n3kiD3
#> 4     5     2 MULTIPOLYGON         2 4u2A8e
#> 5     4     2 MULTIPOLYGON         3 HGSCNS
#>