Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1    1.    2.
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1  3.20  4.00
#> 2  3.00  4.60
#> 3  3.80  4.40
#> 4  3.50  3.80
#> 5  3.40  3.60
#> 6  3.90  4.50
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#> * <dbl> <dbl>
#> 1    0.    3.
#> 2    0.    4.
#> 3    1.    5.
#> 4    2.    5.
#> 
#> $multilinestring
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1 0.     3.00
#>  2 0.     4.00
#>  3 1.00   5.00
#>  4 2.00   5.00
#>  5 0.200  3.00
#>  6 0.200  4.00
#>  7 1.00   4.80
#>  8 2.00   4.80
#>  9 0.     4.40
#> 10 0.600  5.00
#> 
#> $polygon
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1    0.    0.
#>  2    1.    0.
#>  3    3.    2.
#>  4    2.    4.
#>  5    1.    4.
#>  6    0.    0.
#>  7    1.    1.
#>  8    1.    2.
#>  9    2.    2.
#> 10    1.    1.
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1    0.    0.
#>  2    1.    0.
#>  3    3.    2.
#>  4    2.    4.
#>  5    1.    4.
#>  6    0.    0.
#>  7    1.    1.
#>  8    1.    2.
#>  9    2.    2.
#> 10    1.    1.
#> # ... with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_     
#>   <dbl> <dbl> <chr> <chr>     
#> 1    1.    2. POINT b95ecd3932
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     6     2 MULTIPOINT 257e87aef8
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     4     2 LINESTRING 74ad916cbf
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_     
#>   <int> <int> <chr>           <chr>     
#> 1     4     2 MULTILINESTRING 3907083a6d
#> 2     4     2 MULTILINESTRING b656a4f125
#> 3     2     2 MULTILINESTRING 9ed308153c
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_     
#>   <int> <int> <chr>   <chr>     
#> 1     6     2 POLYGON bc2a9bdea8
#> 2     4     2 POLYGON 3154b1a925
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_     
#>   <int> <int> <chr>            <int> <chr>     
#> 1     6     2 MULTIPOLYGON         1 e183c8ab06
#> 2     4     2 MULTIPOLYGON         1 e98c7e1088
#> 3     5     2 MULTIPOLYGON         2 e89dc23320
#> 4     5     2 MULTIPOLYGON         2 cc6d7368f4
#> 5     4     2 MULTIPOLYGON         3 dc0044c630
#> 

## unsure how usefult this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_     
#>   <dbl> <dbl> <chr> <chr>     
#> 1    1.    2. POINT 293f46ba95
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     6     2 MULTIPOINT 1a1f538186
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_     
#>   <int> <int> <chr>      <chr>     
#> 1     4     2 LINESTRING 9aec347fbb
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_     
#>   <int> <int> <chr>           <chr>     
#> 1     4     2 MULTILINESTRING 461b592e63
#> 2     4     2 MULTILINESTRING 812b03580c
#> 3     2     2 MULTILINESTRING fe28bfcf81
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_     
#>   <int> <int> <chr>   <chr>     
#> 1     6     2 POLYGON 64d63c8d2c
#> 2     4     2 POLYGON 67fa663397
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_     
#>   <int> <int> <chr>            <int> <chr>     
#> 1     6     2 MULTIPOLYGON         1 bd2788cc66
#> 2     4     2 MULTIPOLYGON         1 574fbafaf9
#> 3     5     2 MULTIPOLYGON         2 51a7389733
#> 4     5     2 MULTIPOLYGON         2 b995d8205e
#> 5     4     2 MULTIPOLYGON         3 6271c9272d
#>