Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)

#> $point
#> # A tibble: 1 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     1     2
#> 
#> $multipoint
#> # A tibble: 6 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1   3.2   4  
#> 2   3     4.6
#> 3   3.8   4.4
#> 4   3.5   3.8
#> 5   3.4   3.6
#> 6   3.9   4.5
#> 
#> $linestring
#> # A tibble: 4 x 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     0     3
#> 2     0     4
#> 3     1     5
#> 4     2     5
#> 
#> $multilinestring
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1   0     3  
#>  2   0     4  
#>  3   1     5  
#>  4   2     5  
#>  5   0.2   3  
#>  6   0.2   4  
#>  7   1     4.8
#>  8   2     4.8
#>  9   0     4.4
#> 10   0.6   5  
#> 
#> $polygon
#> # A tibble: 10 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> 
#> $multipolygon
#> # A tibble: 24 x 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> # … with 14 more rows
#> 

lapply(sfzoo, sc_path)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT ChOyPQ
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT Jzhkhu
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING QEq7Io
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING t6uBu8
#> 2     4     2 MULTILINESTRING tWr3Qt
#> 3     2     2 MULTILINESTRING d0gFPK
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON I9NM8Y
#> 2     4     2 POLYGON NcQnFu
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 9kRZgW
#> 2     4     2 MULTIPOLYGON         1 HfabxT
#> 3     5     2 MULTIPOLYGON         2 UVdjp2
#> 4     5     2 MULTIPOLYGON         2 1tIUgb
#> 5     4     2 MULTIPOLYGON         3 ETq7ga
#> 

## unsure how usefult this is ...
sc_path(sfgc)

#> $point
#> # A tibble: 1 x 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT 570Qlx
#> 
#> $multipoint
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT Qv0HuS
#> 
#> $linestring
#> # A tibble: 1 x 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING bDGh8k
#> 
#> $multilinestring
#> # A tibble: 3 x 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING GXoOxl
#> 2     4     2 MULTILINESTRING KDMSo8
#> 3     2     2 MULTILINESTRING iW4gSq
#> 
#> $polygon
#> # A tibble: 2 x 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON 7VJQgS
#> 2     4     2 POLYGON obEUKV
#> 
#> $multipolygon
#> # A tibble: 5 x 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 moIAnM
#> 2     4     2 MULTIPOLYGON         1 e67Rj5
#> 3     5     2 MULTIPOLYGON         2 BCoct5
#> 4     5     2 MULTIPOLYGON         2 KnS1vh
#> 5     4     2 MULTIPOLYGON         3 NOcp51
#>