Simple features zoo.

Basic examples of each type of simple feature geometry. sfzoo is a list with each of point, multipoin, linestring, multilinestring, polygon and multipolygon. sfgc is a GEOMETRYCOLLECTION of all the types in sfzoo.

Examples

lapply(sfzoo, sc_coord)
#> $point
#> # A tibble: 1 × 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     1     2
#> 
#> $multipoint
#> # A tibble: 6 × 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1   3.2   4  
#> 2   3     4.6
#> 3   3.8   4.4
#> 4   3.5   3.8
#> 5   3.4   3.6
#> 6   3.9   4.5
#> 
#> $linestring
#> # A tibble: 4 × 2
#>      x_    y_
#>   <dbl> <dbl>
#> 1     0     3
#> 2     0     4
#> 3     1     5
#> 4     2     5
#> 
#> $multilinestring
#>     x_  y_
#> 1  0.0 3.0
#> 2  0.0 4.0
#> 3  1.0 5.0
#> 4  2.0 5.0
#> 5  0.2 3.0
#> 6  0.2 4.0
#> 7  1.0 4.8
#> 8  2.0 4.8
#> 9  0.0 4.4
#> 10 0.6 5.0
#> 
#> $polygon
#>    x_ y_
#> 1   0  0
#> 2   1  0
#> 3   3  2
#> 4   2  4
#> 5   1  4
#> 6   0  0
#> 7   1  1
#> 8   1  2
#> 9   2  2
#> 10  1  1
#> 
#> $multipolygon
#> # A tibble: 24 × 2
#>       x_    y_
#>    <dbl> <dbl>
#>  1     0     0
#>  2     1     0
#>  3     3     2
#>  4     2     4
#>  5     1     4
#>  6     0     0
#>  7     1     1
#>  8     1     2
#>  9     2     2
#> 10     1     1
#> # … with 14 more rows
#> 
lapply(sfzoo, sc_path)
#> $point
#> # A tibble: 1 × 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT 7xNFhx
#> 
#> $multipoint
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT cStxWS
#> 
#> $linestring
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING rxHXgG
#> 
#> $multilinestring
#> # A tibble: 3 × 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING TvSfJl
#> 2     4     2 MULTILINESTRING UdGKCw
#> 3     2     2 MULTILINESTRING KCVaLl
#> 
#> $polygon
#> # A tibble: 2 × 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON UHiMBA
#> 2     4     2 POLYGON lenbah
#> 
#> $multipolygon
#> # A tibble: 5 × 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 cdbkcv
#> 2     4     2 MULTIPOLYGON         1 YFN571
#> 3     5     2 MULTIPOLYGON         2 MeEGFq
#> 4     5     2 MULTIPOLYGON         2 ErR0P4
#> 5     4     2 MULTIPOLYGON         3 SJTIL9
#> 

## unsure how useful this is ...
sc_path(sfgc)
#> $point
#> # A tibble: 1 × 4
#>    nrow  ncol type  path_ 
#>   <dbl> <dbl> <chr> <chr> 
#> 1     1     2 POINT SyBVaI
#> 
#> $multipoint
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     6     2 MULTIPOINT WOABLf
#> 
#> $linestring
#> # A tibble: 1 × 4
#>    nrow  ncol type       path_ 
#>   <int> <int> <chr>      <chr> 
#> 1     4     2 LINESTRING 6DfKms
#> 
#> $multilinestring
#> # A tibble: 3 × 4
#>    nrow  ncol type            path_ 
#>   <int> <int> <chr>           <chr> 
#> 1     4     2 MULTILINESTRING rV5hDx
#> 2     4     2 MULTILINESTRING 5YHoRu
#> 3     2     2 MULTILINESTRING yOL6JC
#> 
#> $polygon
#> # A tibble: 2 × 4
#>    nrow  ncol type    path_ 
#>   <int> <int> <chr>   <chr> 
#> 1     6     2 POLYGON jYzVld
#> 2     4     2 POLYGON T3xTif
#> 
#> $multipolygon
#> # A tibble: 5 × 5
#>    nrow  ncol type         subobject path_ 
#>   <int> <int> <chr>            <int> <chr> 
#> 1     6     2 MULTIPOLYGON         1 CjOs0I
#> 2     4     2 MULTIPOLYGON         1 nKFY6n
#> 3     5     2 MULTIPOLYGON         2 SDGMBO
#> 4     5     2 MULTIPOLYGON         2 xUngYO
#> 5     4     2 MULTIPOLYGON         3 sOZhTm
#>